Put Cache and CacheWrapper in new public header
Summary: The definition of the Cache class should not be needed by the
vast majority of RocksDB users, so I think it is just distracting to
include it in cache.h, which is primarily needed for configuring and
creating caches. This change moves the class to a new header
advanced_cache.h. It is just cut-and-paste except for modifying the
class API comment.

In general, operations on shared_ptr<Cache> should continue to work when
only a forward declaration of Cache is available, as long as all the
Cache instances provided are already shared_ptr. See
https://stackoverflow.com/a/17650101/454544

Also, the most common way to customize a Cache is by wrapping an
existing implementation, so it makes sense to provide CacheWrapper in
the public API. This was a cut-and-paste job except removing the
implementation of Name() so that derived classes must provide it.

Intended follow-up: consolidate Release() into one function to reduce
customization bugs / confusion

Test Plan: `make check`
