+++ utilities/simulator_cache/sim_cache.cc
@@ -151,20 +151,22 @@ class SimCacheImpl : public SimCache {
   // capacity for real cache (ShardedLRUCache)
   // test_capacity for key only cache
   SimCacheImpl(std::shared_ptr<Cache> sim_cache, std::shared_ptr<Cache> cache)
-      : cache_(cache),
+      : SimCache(cache),
         key_only_cache_(sim_cache),
         miss_times_(0),
         hit_times_(0),
         stats_(nullptr) {}
 
   ~SimCacheImpl() override {}
-  void SetCapacity(size_t capacity) override { cache_->SetCapacity(capacity); }
+
+  const char* Name() const override { return "SimCache"; }
+
+  void SetCapacity(size_t capacity) override { target_->SetCapacity(capacity); }
 
   void SetStrictCapacityLimit(bool strict_capacity_limit) override {
-    cache_->SetStrictCapacityLimit(strict_capacity_limit);
+    target_->SetStrictCapacityLimit(strict_capacity_limit);
   }
 
-  using Cache::Insert;
   Status Insert(const Slice& key, Cache::ObjectPtr value,
                 const CacheItemHelper* helper, size_t charge, Handle** handle,
                 Priority priority) override {
@@ -184,10 +186,10 @@ class SimCacheImpl : public SimCache {
     }
 
     cache_activity_logger_.ReportAdd(key, charge);
-    if (!cache_) {
+    if (!target_) {
       return Status::OK();
     }
-    return cache_->Insert(key, value, helper, charge, handle, priority);
+    return target_->Insert(key, value, helper, charge, handle, priority);
   }
 
   Handle* Lookup(const Slice& key, const CacheItemHelper* helper,
@@ -195,54 +197,54 @@ class SimCacheImpl : public SimCache {
                  Priority priority = Priority::LOW, bool wait = true,
                  Statistics* stats = nullptr) override {
     HandleLookup(key, stats);
-    if (!cache_) {
+    if (!target_) {
       return nullptr;
     }
-    return cache_->Lookup(key, helper, create_context, priority, wait, stats);
+    return target_->Lookup(key, helper, create_context, priority, wait, stats);
   }
 
-  bool Ref(Handle* handle) override { return cache_->Ref(handle); }
+  bool Ref(Handle* handle) override { return target_->Ref(handle); }
 
   using Cache::Release;
   bool Release(Handle* handle, bool erase_if_last_ref = false) override {
-    return cache_->Release(handle, erase_if_last_ref);
+    return target_->Release(handle, erase_if_last_ref);
   }
 
   void Erase(const Slice& key) override {
-    cache_->Erase(key);
+    target_->Erase(key);
     key_only_cache_->Erase(key);
   }
 
   Cache::ObjectPtr Value(Handle* handle) override {
-    return cache_->Value(handle);
+    return target_->Value(handle);
   }
 
-  uint64_t NewId() override { return cache_->NewId(); }
+  uint64_t NewId() override { return target_->NewId(); }
 
-  size_t GetCapacity() const override { return cache_->GetCapacity(); }
+  size_t GetCapacity() const override { return target_->GetCapacity(); }
 
   bool HasStrictCapacityLimit() const override {
-    return cache_->HasStrictCapacityLimit();
+    return target_->HasStrictCapacityLimit();
   }
 
-  size_t GetUsage() const override { return cache_->GetUsage(); }
+  size_t GetUsage() const override { return target_->GetUsage(); }
 
   size_t GetUsage(Handle* handle) const override {
-    return cache_->GetUsage(handle);
+    return target_->GetUsage(handle);
   }
 
   size_t GetCharge(Handle* handle) const override {
-    return cache_->GetCharge(handle);
+    return target_->GetCharge(handle);
   }
 
   const CacheItemHelper* GetCacheItemHelper(Handle* handle) const override {
-    return cache_->GetCacheItemHelper(handle);
+    return target_->GetCacheItemHelper(handle);
   }
 
-  size_t GetPinnedUsage() const override { return cache_->GetPinnedUsage(); }
+  size_t GetPinnedUsage() const override { return target_->GetPinnedUsage(); }
 
   void DisownData() override {
-    cache_->DisownData();
+    target_->DisownData();
     key_only_cache_->DisownData();
   }
 
@@ -250,11 +252,11 @@ class SimCacheImpl : public SimCache {
       const std::function<void(const Slice& key, ObjectPtr value, size_t charge,
                                const CacheItemHelper* helper)>& callback,
       const ApplyToAllEntriesOptions& opts) override {
-    cache_->ApplyToAllEntries(callback, opts);
+    target_->ApplyToAllEntries(callback, opts);
   }
 
   void EraseUnRefEntries() override {
-    cache_->EraseUnRefEntries();
+    target_->EraseUnRefEntries();
     key_only_cache_->EraseUnRefEntries();
   }
 
@@ -295,7 +297,7 @@ class SimCacheImpl : public SimCache {
   std::string GetPrintableOptions() const override {
     std::ostringstream oss;
     oss << "    cache_options:" << std::endl;
-    oss << cache_->GetPrintableOptions();
+    oss << target_->GetPrintableOptions();
     oss << "    sim_cache_options:" << std::endl;
     oss << key_only_cache_->GetPrintableOptions();
     return oss.str();
@@ -314,7 +316,6 @@ class SimCacheImpl : public SimCache {
   }
 
  private:
-  std::shared_ptr<Cache> cache_;
   std::shared_ptr<Cache> key_only_cache_;
   std::atomic<uint64_t> miss_times_;
   std::atomic<uint64_t> hit_times_;
